backend/requirements.txt
django
djangorestframework
strawberry-graphql
django-cors-headers
python-dotenv
psycopg2-binary
dj-database-url
pika==1.3.2
requests==2.32.3
celery==5.4.0
flower==2.0.1